Finishing up refactoring
